fixed a bug in __init__.py that prevente dscipy.weave.test() from working.
various clean up and changes to testing and to catalog pointed out by Prabhu on his Debian machine.  Hopefully all tests will pass for him now.

Added a try/except to prevent loading of corupted catalogs in add_persistent.  I can't imagine why a catalog would get corrupted, but it was happening to Prabhu.  Whether this is fixing a symptom of something else, or is really a good fix remains to be seen.