fixed bug where exceptions during initializing variables could cause problems with during cleanup of variables that were not initialized. Now a flag is set for each variable when it is initialized and the clean up code is only executed for variables that have been initialized.
A init_flag() variable was added to the base_spec to help with the
above bug fix.

The SWIG_GetPtr method does some sophisticated type checking that
weave really doesn't need.  Most of this has been commented out
and a simple string comparison is done to make sure types are correct.