Last month, I made some changes so that these modules would be built in the build/os-xxx.temp directory instead of cluttering the directory where setup.py lives.  This then requires the compiler to be told where the module files live.  The inclusion of the temp directory should be done automatically because the module files are placed there automatically.
Changes over the last month (I think in revision 1.58) broke this.  It looks like it was to fix incompatibilities for other absoft compilers.  Dang they change their switches a lot...  Anyway, all absoft f90 compilers should allow a way of specifying the modules directory.  Instead of disabling the search of the temp directory, we should find out how the other versions do it.

I've re-enabled the search for modules in the temp direcotry for the time being.