Implemented hooks to postpone module import until module attributes are accessed (ppimport function).  To model postponed import for 'from module import name', ppimport_attr(module,name) is introduced where module=ppimport(modulename).
Comparing ppimport to Pmw dynamic loader, ppimport is more robust: it
does not assume any special setup for a module to be imported. With
ppimport one can postpone importing both python modules and packages
as well as python extension modules. ppimport covers the two first
purposes of a dynamic loader (as defined by Pmw docs). The third
purpose, that is choosing between different versions of a module,
should be implemented separately (if at all at this moment) from
ppimport to keep its code as clean as possible.

The basic idea of ppimport is as follows. Expression

  module = ppimport('module')

is equivalent to

  import module

with the difference that the actual module import is postponed until
any part of a python code tries to access the attributes of a module.
The return value of ppimport is either actual module (when module
was imported before ppimport call) or a special ModuleLoader instance
(that triggers module import as soon as module attributes are to be
accessed either through getattr or setattr).

Restrictions: ModuleLoader is not derived from Python module type.
As a result, reload(module) will raise TypeError. As a workaround,
reload(module._ppimport_module) can be used.

Wishes: currently module.__doc__ would trigger importing module though
it may not be necessary. Therefore, a database of modules
documentation is needed in some form (e.g. definition files per
package, preferably generated automatically during build/installation
steps) to prevent unnecessary imports. We need to work out a policy
for documenting scipy modules... what tools to use (pydoc, etc.),
documentation string formats, etc.