Re #1008: add terms to Horner series for long doubles. Fix issues in test.
The Horner series was previously too short to accommodate SPARC quad
precision. There were also some issues with tests on platforms with
bogus longdouble -- finfo would report invalid values for eps/tiny.
Try to detect this now, and to be sure, use only limited magnitude
range for test points (this is sufficient for testing that the series
expansion works).