ENH: configure NPY_INLINE when the header is read, not at build time.
We configured the macro NPY_INLINE (for portable inline) at numpy build
time, but this problematic when a numpy extension is built with a
different compiler than the one used to build numpy itsel (e.g. building
numpy with mingw, but building a numpy extension with MS compiler).
Instead, the macro is defined everytime npy_common.h is read.